+++ ExtrasKitDemo-iOS/ExtrasKitDemo-iOS/RootViewController.swift
@@ -23,8 +23,13 @@ class RootViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = "bbbbbb".ek.color
-        var asd = [1, 2, 9]
-        let result = asd.ek.intersection([1, 2, 3], [2, 3, 4], [2, 4, 5, 6]) { $0 }
-        
+//        var asd = [1, 2, 9]
+//        var result = asd.ek.intersection([1, 2, 3], [2, 3, 4], [2, 4, 5, 6]) { $0 }
+//        print(asd)
+//        print(result)
+
+        var asd = [1, 1, 2, 3]
+        print(asd.ek.union([2, 3, 4], where: { $0 }))
+        print(asd)
     }
 }