+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add a method that create a file URL referencing the local file or directory at path.
 - Add a method that diff the original and the given sequences.
 - Add a method that intersect the original and the given sequences.
+- Add a method that crop image within aspect ratio specified by a CGSize.
 
 ### Fixed
 <for any bug fixes.>