+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add a method that diff the original and the given sequences.
 - Add a method that intersect the original and the given sequences.
 - Add a method that crop image within aspect ratio specified by a CGSize.
+- Add a method that union the original and the given sequences.
 
 ### Fixed
 <for any bug fixes.>